feat(ensrainbow): introduce `ensrainbow-sdk`
Uses types and consts from a package that is dedicated for both, client and server applications.
