+++ apps/ensrainbow/src/index.test.ts
@@ -1,12 +1,12 @@
 import { serve } from "@hono/node-server";
+import { ErrorCode, StatusCode } from "ensrainbow-sdk/consts";
+import type { CountResponse, HealError, HealResponse, HealSuccess } from "ensrainbow-sdk/types";
 import { labelhash } from "viem";
 /// <reference types="vitest" />
 import { afterAll, beforeAll, beforeEach, describe, expect, it } from "vitest";
 import { app, db } from "./index";
 import { LABELHASH_COUNT_KEY } from "./utils/constants";
 import { labelHashToBytes } from "./utils/label-utils";
-import type { CountResponse, HealError, HealResponse, HealSuccess } from "./utils/response-types";
-import { ErrorCode, StatusCode } from "./utils/response-types";
 
 describe("ENS Rainbow API", () => {
   let server: ReturnType<typeof serve>;
