+++ apps/ensrainbow/src/operations.ts
@@ -1,7 +1,5 @@
 import { ClassicLevel } from "classic-level";
-import { ByteArray } from "viem";
-import { LABELHASH_COUNT_KEY } from "./utils/constants.js";
-import { labelHashToBytes } from "./utils/label-utils.js";
+import { ErrorCode, StatusCode } from "ensrainbow-sdk/consts";
 import type {
   CountError,
   CountResponse,
@@ -9,8 +7,10 @@ import type {
   HealError,
   HealResponse,
   HealSuccess,
-} from "./utils/response-types.js";
-import { ErrorCode, StatusCode } from "./utils/response-types.js";
+} from "ensrainbow-sdk/types";
+import { ByteArray } from "viem";
+import { LABELHASH_COUNT_KEY } from "./utils/constants.js";
+import { labelHashToBytes } from "./utils/label-utils.js";
 
 export interface ENSRainbowContext {
   db: ClassicLevel<ByteArray, string>;
