+++ packages/ensrainbow-sdk/package.json
@@ -10,8 +10,13 @@
     "directory": "packages/ensrainbow-sdk"
   },
   "homepage": "https://github.com/namehash/ensnode/tree/main/packages/ensrainbow-sdk",
+  "exports": {
+    "./consts": "./src/consts.ts",
+    "./types": "./src/types.ts"
+  },
   "scripts": {
     "typecheck": "tsc --noEmit",
+    "lint": "biome check --write",
     "lint:ci": "biome ci"
   },
   "devDependencies": {
